# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/rol/{id}": {
      "get": {
        "description": "Retrieves a role by its unique identifier",
        "operationId": "RolController_getRolById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Role ID",
            "schema": {
              "example": "rol-123"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Role found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Rol"
                }
              }
            }
          },
          "404": {
            "description": "Role not found"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get role by ID",
        "tags": [
          "rol"
        ]
      }
    },
...
    "/suggestion/{id}": {
      "get": {
        "description": "Retrieves a suggestion by its unique identifier",
        "operationId": "SuggestionController_getSuggestionById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Suggestion ID",
            "schema": {
              "example": "suggestion-123"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Suggestion found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Suggestion"
                }
              }
            }
          },
          "404": {
            "description": "Suggestion not found"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get suggestion by ID",
        "tags": [
          "suggestion"
        ]
      }
    },
    "/suggestion": {
      "post": {
        "description": "Creates a new user suggestion or feedback (public endpoint)",
        "operationId": "SuggestionController_createSuggestion",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "FEATURE_REQUEST"
                  },
                  "comment": {
                    "type": "string",
                    "example": "User feedback comment"
                  },
                  "userId": {
                    "type": "string",
                    "example": "user-123"
                  },
                  "userType": {
                    "type": "string",
                    "example": "CLIENT"
                  }
                },
                "required": [
                  "type",
                  "comment"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Suggestion created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Suggestion"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          }
        },
        "summary": "Create a new suggestion",
        "tags": [
          "suggestion"
        ]
      }
    },
...
  },
  "info": {
    "title": "ESV Backend API",
    "description": "API documentation for ESV Backend - Appointment and Queue Management System",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "https://app.easuavez.com",
      "description": "Default server"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "Rol": {
        "type": "object",
        "properties": {}
      },
...
      "Suggestion": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}